Create Lc2AddTwoNumbersII.h
Push nodes into a stack and then pop them one by one, we get the least significant digits first. Then we can add numbers as normal. The result nodes push to a stack, and then we read them again, to build the result list from header to tail. So we get the most significant digits first.